backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.10.4
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.4.0
gensim==4.3.3
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
libclang==18.1.1
LibRecommender==1.5.1
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pinecone==5.1.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scikit_learn==1.5.2
pytest==8.3.3
httpx==0.27.2
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
starlette==0.40.0
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
